+++ project_4_web_service/src/main/java/io/ab/library/repository/PublisherRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import io.ab.library.model.Publisher;
 
-public interface PublisherRepository extends CrudRepository<Publisher, Long> {
+public interface PublisherRepository extends CrudRepository<Publisher, Integer> {
 
 }
