+++ project_4_web_service/src/main/java/io/ab/library/repository/TagRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import io.ab.library.model.Tag;
 
-public interface TagRepository extends CrudRepository<Tag, Long> {
+public interface TagRepository extends CrudRepository<Tag, Integer> {
 
 }
